Split apart BrokerClient tests into separate files
