+++ alpaca/broker/client.py
@@ -587,3 +587,5 @@ class BrokerClient(RESTClient):
             # receive them from the api
             for chunk in response.iter_content(chunk_size=None):
                 f.write(chunk)
+
+    # ############################## FUNDING ################################# #
