+++ tests/broker/broker_client/test_documents_routes.py
@@ -1,0 +1,266 @@
+import os.path
+import tempfile
+from typing import Iterator, List
+
+import pytest
+import requests_mock
+from requests_mock import Mocker
+
+from alpaca.broker import UploadDocumentRequest
+from alpaca.broker.client import BrokerClient
+from alpaca.broker.constants import BROKER_DOCUMENT_UPLOAD_LIMIT
+from alpaca.broker.enums import (
+    TradeDocumentType,
+    UploadDocumentMimeType,
+    UploadDocumentType,
+)
+from alpaca.broker.models import GetTradeDocumentsRequest, TradeDocument
+from alpaca.common.enums import BaseURL
+
+
+@pytest.fixture
+def reqmock() -> Iterator[Mocker]:
+    with requests_mock.Mocker() as m:
+        yield m
+
+
+@pytest.fixture
+def client():
+    client = BrokerClient(
+        "key-id",
+        "secret-key",
+        sandbox=True,  # Expressly call out sandbox as true for correct urls in reqmock
+    )
+    return client
+
+
+@pytest.fixture
+def raw_client():
+    raw_client = BrokerClient("key-id", "secret-key", raw_data=True)
+    return raw_client
+
+
+def test_get_trade_documents_for_account(reqmock, client: BrokerClient):
+    account_id = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+
+    reqmock.get(
+        BaseURL.BROKER_SANDBOX + f"/v1/accounts/{account_id}/documents",
+        text="""
+        [
+          {
+            "id": "1b560b0f-9efd-44b4-8004-dfd520c7cdc0",
+            "name": "",
+            "type": "account_statement",
+            "sub_type": "",
+            "date": "2022-02-27"
+          },
+          {
+            "id": "c2619f26-4cb3-4ef2-8c4d-660dc561b42d",
+            "name": "",
+            "type": "account_statement",
+            "sub_type": "",
+            "date": "2022-02-28"
+          }
+        ]
+        """,
+    )
+
+    start = "2022-02-27"
+    end = "2022-02-28"
+
+    documents = client.get_trade_documents_for_account(
+        account_id=account_id,
+        documents_filter=GetTradeDocumentsRequest(
+            start=start, end=end, type=TradeDocumentType.ACCOUNT_STATEMENT
+        ),
+    )
+
+    assert reqmock.called_once
+    assert isinstance(documents, List)
+    assert isinstance(documents[0], TradeDocument)
+
+    assert reqmock.request_history[0].qs == {
+        "start": [start],
+        "end": [end],
+        "type": [TradeDocumentType.ACCOUNT_STATEMENT],
+    }
+
+
+def test_get_trade_documents_for_account_validates_account_id(
+    reqmock,
+    client: BrokerClient,
+):
+    with pytest.raises(ValueError):
+        client.get_trade_documents_for_account(
+            account_id="not a uuid", documents_filter=GetTradeDocumentsRequest()
+        )
+
+
+def test_upload_documents_to_account(reqmock, client: BrokerClient):
+    account_id = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+    reqmock.post(
+        BaseURL.BROKER_SANDBOX + f"/v1/accounts/{account_id}/documents/upload",
+        json={},
+        status_code=202,
+    )
+
+    client.upload_documents_to_account(
+        account_id=account_id,
+        document_data=[
+            UploadDocumentRequest(
+                document_type=UploadDocumentType.ACCOUNT_APPROVAL_LETTER,
+                content="fake base64",
+                mime_type=UploadDocumentMimeType.PDF,
+            )
+        ],
+    )
+
+    assert reqmock.called_once
+
+    # TODO: Add a custom reqmock matcher to ensure format of request rather than this static string check
+    assert (
+        reqmock.request_history[0].text
+        == '[{"document_type": "account_approval_letter", "content": "fake base64", "mime_type": "application/pdf"}]'
+    )
+
+
+def test_upload_documents_to_account_validates_limit(reqmock, client: BrokerClient):
+    with pytest.raises(ValueError) as e:
+        client.upload_documents_to_account(
+            account_id="2a87c088-ffb6-472b-a4a3-cd9305c8605c",
+            document_data=list(range(0, 11)),
+        )
+
+    assert not reqmock.called
+    assert f"document_data cannot be longer than {BROKER_DOCUMENT_UPLOAD_LIMIT}" in str(
+        e.value
+    )
+
+
+def test_upload_documents_to_account_validates_account_id(
+    reqmock,
+    client: BrokerClient,
+):
+    with pytest.raises(ValueError):
+        client.upload_documents_to_account(
+            account_id="not a uuid", document_data=list(range(1, 10))
+        )
+
+    assert not reqmock.called
+
+
+def test_get_trade_document_for_account_by_id_validates_uuids(
+    reqmock,
+    client: BrokerClient,
+):
+    uuid = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+
+    with pytest.raises(ValueError):
+        client.get_trade_document_for_account_by_id(
+            account_id=uuid, document_id="not a uuid"
+        )
+
+    with pytest.raises(ValueError):
+        client.get_trade_document_for_account_by_id(
+            account_id="not a uuid",
+            document_id=uuid,
+        )
+
+
+def test_get_trade_document_for_account_by_id(reqmock, client: BrokerClient):
+    account_id = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+    document_id = "2a87c089-ffb6-472b-a4a3-cd9305c8605d"
+
+    reqmock.get(
+        BaseURL.BROKER_SANDBOX + f"/v1/accounts/{account_id}/documents/{document_id}",
+        text="""
+        {
+          "id": "1b560b0f-9efd-44b4-8004-dfd520c7cdc0",
+          "name": "",
+          "type": "account_statement",
+          "sub_type": "",
+          "date": "2022-02-28"
+        }
+        """,
+    )
+
+    result = client.get_trade_document_for_account_by_id(
+        account_id=account_id,
+        document_id=document_id,
+    )
+
+    assert reqmock.called_once
+    assert isinstance(result, TradeDocument)
+
+
+def test_download_trade_document_for_account_by_id(reqmock, client: BrokerClient):
+    """
+    High level steps for how to do this test:
+      * make a tempdir as with for auto delete
+      * get a tempfile name in said tempdir
+      * mock out main api request that returns a fake url
+      * mock out fake url with fake data
+      * check tempfile to ensure size is as expected
+        * this could technically have issues with different encodings n such
+        * maybe just read back out the data to ensure it's the same
+
+    going to keep this docblock here to show thought process/possible downsides for future ref if this test proves
+    too flakey
+    """
+    account_id = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+    document_id = "2a87c089-ffb6-472b-a4a3-cd9305c8605d"
+    fake_dl_url = f"https://fake-dl.com/{document_id}"
+    fake_dl_text = """
+    here is
+    some fake
+    text
+    """
+
+    reqmock.get(
+        BaseURL.BROKER_SANDBOX
+        + f"/v1/accounts/{account_id}/documents/{document_id}/download",
+        status_code=301,
+        headers={"Location": fake_dl_url},
+        text=f"""
+        <a href="{fake_dl_url}">
+          Moved Permanently
+        </a>.
+        """,
+    )
+
+    reqmock.get(fake_dl_url, text=fake_dl_text)
+
+    with tempfile.TemporaryDirectory() as tempdir:
+        tempname = os.path.join(tempdir, "test.pdf")
+        print(tempname)
+
+        client.download_trade_document_for_account_by_id(
+            account_id=account_id, document_id=document_id, file_name=tempname
+        )
+
+        assert reqmock.call_count == 2
+
+        assert os.path.isfile(tempname)
+
+        with open(tempname, mode="r") as file:
+            assert file.read() == fake_dl_text
+
+
+def test_download_trade_document_for_account_by_id_validates_uuids(
+    reqmock, client: BrokerClient
+):
+    uuid = "2a87c088-ffb6-472b-a4a3-cd9305c8605c"
+
+    with pytest.raises(ValueError):
+        client.download_trade_document_for_account_by_id(
+            account_id=uuid,
+            document_id="not a uuid",
+            file_name="",
+        )
+
+    with pytest.raises(ValueError):
+        client.download_trade_document_for_account_by_id(
+            account_id="not a uuid",
+            document_id=uuid,
+            file_name="",
+        )
